+++ templates/home.html
@@ -24,15 +24,16 @@
 
 	{% endif %}
 
-	<div class="text-center" id="mainpage">
-		<h3>What are my options?</h3>
-		<div class="btn-group">
-			<a href="/books" class="btn btn-primary btn-md">My Books</a>
+	<div class="jumbotron text-center" id="mainpage">
+		<div class="text-center" id="whiteplease">
+			<h3>What are my options?</h3>
+			<div class="btn-group">
+				<a href="/books" class="btn btn-primary btn-md">My Books</a>
 
-			<a href="/add_book" class="btn btn-info btn-md">Register A Book</a>
+				<a href="/add_book" class="btn btn-info btn-md">Register A Book</a>
 
-			<a href="/profile" class="btn btn-info btn-md">My Profile</a>
+				<a href="/profile" class="btn btn-info btn-md">My Profile</a>
+			</div>
 		</div>
 	</div>
-
 {% endblock %}
