+++ templates/layout.html
@@ -12,8 +12,8 @@
 	<div class="container">
 		{% include 'includes/_messages.html' %}
 		{% block body%}{%endblock%}
-		
-	</div> 
+	</div>
+	
 	<script src = "https://stackpath.bootstrapcdn.com/bootstrap/4.1.2/js/bootstrap.min.js"></script>
 	<script src="//cdn.ckeditor.com/4.6.2/basic/ckeditor.js"></script>
 	<script type="text/javascript">
